build/CI, changed files that say they have no changes?
Reasonable gradle jvm args.
Remove travis config, we don't use this.
Up gradle version to 5.6.1.

Fixed dependency configurations, running the tests requires the test dependencies to be in the test runtime classpath along with the test compile classpath.
Add extra spotbugs checks via spotbugsPlugins dependencies, sb-contrib and findsecbugs.

Update URL to map to forked repo.

Add Eclipse plugin with eclipse spotbugs configuration.

Fix spotbugs plugin resolution.
Make task name align with artifact classifier.
Add java-library plugin since this is a java library and can benefit from the conventions that plugin applies, remove bintray plugin since this fork should not be published to bintray, and add Eclipse and maven-publish plugins to ensure it all works as expected and there are no missing build dependencies.

For this fork, we are not publishing to bintray, so configuration for that is removed.